requirements.txt
Flask
Flask-WTF
Flask-Login
Flask-PyMongo
WTForms
python-dotenv
dnspython
